Refactor to export testutils (#207)
* refactor to export testutils

* linter
